chang type lat lng in orders entity

+++ src/main/java/com/example/backend/model/orderModel/OrderReq.java
@@ -23,9 +23,9 @@ public class OrderReq {
 
     private String detail;
 
-    private Long lat;
+    private Float lat;
 
-    private Long lng;
+    private Float lng;
 
     private String address;
